==> Gene expression Analysis/Annotation/03_T&NK_annotation.R <==
library(Seurat)
library(harmony)
library(tidyverse)
library(ggplot2)
library(cowplot)
library(dplyr)
library(patchwork)
library(R.utils)
.cluster_cols <- c("#DC050C", "#FB8072", "#1965B0", "#7BAFDE", "#882E72",
                   "#B17BA6", "#FF7F00", "#FDB462", "#E7298A", "#E78AC3",
                   "#33A02C", "#B2DF8A", "#55A1B1", "#8DD3C7", "#A6761D",
                   "#E6AB02", "#7570B3", "#BEAED4", "#666666", "#999999",
                   "#aa8282", "#d4b7b7", "#8600bf", "#ba5ce3", "#808000",
                   "#aeae5c", "#1e90ff", "#00bfff", "#56ff0d", "#ffff00")

# define a function for sub-clustering
subcluster <- function(seurat_obj, dims_use, resolution, n_neigh = 30, min_dist = 0.3) {
  seurat_obj <- NormalizeData(seurat_obj)
  seurat_obj <- FindVariableFeatures(seurat_obj, selection.method = "vst", nfeatures = 2000)
  seurat_obj <- ScaleData(seurat_obj)
  seurat_obj <- RunPCA(seurat_obj, features = VariableFeatures(seurat_obj))
  seurat_obj <- RunHarmony(seurat_obj, group.by.vars = "orig.ident", max_iter = 30) # nolint
  seurat_obj <- FindNeighbors(seurat_obj, reduction = "harmony", dims = dims_use)
  seurat_obj <- FindClusters(seurat_obj, resolution = resolution)
  seurat_obj <- RunUMAP(seurat_obj, dims = dims_use, reduction = "harmony",
                        n.neighbors = n_neigh, min.dist = min_dist)
  return(seurat_obj)
}
# L1 annotation for T&NK cells
pbmc <- readRDS("/data/RDS/Annotation/3_PBMC_annotation_L1.rds")
sub_data <- subset(pbmc, subset = celltype == "T&NK")
sub_data <- subcluster(sub_data, dims_use = 1:7, resolution = 1.2)
T_NK_marker <- c("CD3D","CD3G","CD4","CD8A","CD8B","NCAM1","FCGR3A","NCR1","IL7R","KLRC1","KLRD1","CD160","XCL1","MKI67")
DimPlot(sub_data, label = TRUE, cols = .cluster_cols, pt.size = 0.5)
FeaturePlot(sub_data, features = T_NK_marker, ncol = 4)
sub_data <- RenameIdents(sub_data,
                         '0' = "CD8 T & other T", '1' = "NK", '2' = "NK", '3' = "NK", '4' = "NK", 
                         '5' = "CD8 T & other T", '6' = "CD8 T & other T", '7' = "CD8 T & other T", '8' = "CD8 T & other T",
                         '9' = "CD8 T & other T", '10' = "NK", '11' = "CD8 T & other T", '12' = "NK",
                         '13' = "CD8 T & other T", '14' = "NK", '15' = "CD8 T & other T")
sub_data$celltype <- Idents(sub_data)
pbmc$celltype[colnames(sub_data)] <- sub_data$celltype
#NK subclustering
NK <- subset(sub_data, subset = celltype == "NK")
NK <- subcluster(NK, dims_use = 1:7, resolution = 1.2)
FeaturePlot(NK, features = T_NK_marker, ncol = 4)
NK <- RenameIdents(NK,
                       '0' = "NK", '1' = "NK", '2' = "CD8 T & other T", '3' = "NK",
                       '4' = "NK", '5' = "NK", '6' = "NK", '7' = "CD8 T & other T",
                       '8' = "CD8 T & other T", '9' = "NK", '10' = "NK",
                       '11' = "NK", '12' = "CD8 T & other T", '13' = "NK")
NK$celltype <- Idents(NK)
pbmc$celltype[colnames(NK)] <- NK$celltype
#CD8 T subclustering
CD8_T <- subset(sub_data, subset = celltype == "CD8 T & other T")
CD8_T <- subcluster(CD8_T, dims_use = 1:8, resolution = 1.0)
L2_CD8_marker <- c("TRABD2A","MAL","CCR7","SELL","LEF1","TCF7","GPR183","CD44","S100A4",
                   "GZMA","GZMH","NKG7","GNLY","MKI67","CABP4","NOTCH1","TRDC","SLC4A10",
                   "KLRB1","KLRF1","KLRC2","TYROBP","FCGR3A")
FeaturePlot(CD8_T, features = L2_CD8_marker, ncol = 4)
CD8_T <- RenameIdents(CD8_T,
                      '0' = "NKT",'1' = "CD8 T & other T", '2' = "NKT", '3' = "CD8 T & other T", 
                      '4' = "CD8 T & other T", '5' = "CD8 Naïve", '6' = "CD8 Naïve", 
                      '7' = "MAIT",'8' = "CD8 Memory",'9' = "CD8 T & other T",'10' = "CD8 Memory",
                      '11' = "unknown",'12' = "Cycling T")
CD8_T$L2_celltype <- Idents(CD8_T)
pbmc$celltype[colnames(CD8_T)] <- CD8_T$L2_celltype
#CTL subclustering
CD8_T_main <- subset(CD8_T, subset = L2_celltype == "CD8 T & other T")
CD8_T_main <- subcluster(CD8_T_main, dims_use = 1:8, resolution = 1.0)
FeaturePlot(CD8_T_main, features = c("CD3D","CD4","CD8A","CD8B"), ncol = 4)
CD8_T_main <- RenameIdents(CD8_T_main,
                           '0' = "CD8 T & other T",'1' = "CD8 T & other T", '2' = "CD8 T & other T", 
                           '3' = "CD8 T & other T",'4' = "CD8 T & other T",'5' = "CD8 T & other T",
                           '6' = "CD8 T & other T",'7' = "CD4 T",'8' = "CD8 T & other T",
                           '9' = "CD8 T & other T",'10' = "CD4 T")
pbmc$celltype[colnames(CD8_T_main)] <- Idents(CD8_T_main)

saveRDS(pbmc, "/data/RDS/Annotation/4_PBMC_annotation_L1.rds")
