==> Gene expression Analysis/Annotation/05_CD4T_annotation.R <==
library(Seurat)
library(harmony)
library(tidyverse)
library(ggplot2)
library(cowplot)
library(dplyr)
library(patchwork)
library(R.utils)
.cluster_cols <- c("#DC050C", "#FB8072", "#1965B0", "#7BAFDE", "#882E72",
                   "#B17BA6", "#FF7F00", "#FDB462", "#E7298A", "#E78AC3",
                   "#33A02C", "#B2DF8A", "#55A1B1", "#8DD3C7", "#A6761D",
                   "#E6AB02", "#7570B3", "#BEAED4", "#666666", "#999999",
                   "#aa8282", "#d4b7b7", "#8600bf", "#ba5ce3", "#808000",
                   "#aeae5c", "#1e90ff", "#00bfff", "#56ff0d", "#ffff00")

# Function for sub-clustering
subcluster <- function(seurat_obj, dims_use, resolution, nfeatures = 2000, n_neigh = 30, min_dist = 0.3) {
  seurat_obj <- NormalizeData(seurat_obj)
  seurat_obj <- FindVariableFeatures(seurat_obj, selection.method = "vst", nfeatures = nfeatures)
  seurat_obj <- ScaleData(seurat_obj)
  seurat_obj <- RunPCA(seurat_obj, features = VariableFeatures(seurat_obj))
  seurat_obj <- RunHarmony(seurat_obj, group.by.vars = "orig.ident", max_iter = 30) # nolint
  seurat_obj <- FindNeighbors(seurat_obj, reduction = "harmony", dims = dims_use)
  seurat_obj <- FindClusters(seurat_obj, resolution = resolution)
  seurat_obj <- RunUMAP(seurat_obj, dims = dims_use, reduction = "harmony",
                        n.neighbors = n_neigh, min.dist = min_dist)
  return(seurat_obj)
}
#L2 CD4T annotation
all_cell_data <- readRDS("/data/RDS/Annotation/4_PBMC_annotation_L1.rds")
CD4_T <- subset(all_cell_data, subset = celltype == "CD4 T")
CD4_T <- subcluster(CD4_T, dims_use = 1:11, resolution = 0.8)
CD4_T_marker <- c("TRABD2A", "MAL", "CCR7", "SELL", "LEF1", "TCF7", "GPR183",
                  "CD44", "S100A4", "CTLA4", "IL7R", "CX3CR1", "GZMA", "GZMH", "NKG7", "GNLY")
FeaturePlot(CD4_T, features = CD4_T_marker, ncol = 4)
CD4_T <- RenameIdents(CD4_T,
                      '0' = "CD4 Naïve", '1' = "CD4 Naïve",
                      '2' = "CD4 Memory", '3' = "CD4 Naïve",
                      '4' = "CD4 Memory", '5' = "CD4 Naïve",
                      '6' = "CD4 Memory", '7' = "CD4 Treg",
                      '8' = "CD4 CTL", '9' = "CD4 Memory")
CD4_T$L3_celltype <- Idents(CD4_T)
#L3 CD4T annotation
CD4_Memory <- subset(CD4_T, subset = L3_celltype == "CD4 Memory")
CD4_Memory <- subcluster(CD4_Memory, dims_use = 1:11, resolution = 0.8)
options(repr.plot.height = 6, repr.plot.width = 6)
FeaturePlot(CD4_Memory, features = "CCR7")
DotPlot(CD4_Memory, features = "CCR7", assay = "RNA", cols = "RdBu")
CD4_Memory <- RenameIdents(CD4_Memory,
                           '0' = "CD4 Tcm", '1' = "CD4 Tcm",
                           '2' = "CD4 Tem", '3' = "CD4 Tem",
                           '4' = "CD4 Tem", '5' = "CD4 Tem",
                           '6' = "CD4 Tem")
CD4_Memory$L3_celltype <- Idents(CD4_Memory)
CD4_T$L3_celltype[colnames(CD4_Memory)] <- CD4_Memory$L3_celltype
saveRDS(CD4_T, "/data/RDS/Annotation/6_CD4T_annotation.rds")
